refactor: separate login from router
