+++ src/components/HeaderUser/HeaderUser.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { UserContext } from '../../App';
+import { UserContext } from '../../pages/HomePage/HomePage';
 
 interface HeaderUserProps extends React.HTMLAttributes<HTMLElement> {
   src?: string;
